refactor: move shared get_token method to utils
