+++ src/dexes/uniswap_v2.rs
@@ -85,19 +85,6 @@ impl DexProtocol for UniswapV2 {
         Ok(Vec::new())
     }
 
-    async fn get_token(&self, token_address: Address) -> Result<Token, Error> {
-        // This is a placeholder implementation
-        // In production, we'd use provider.call() with correct parameters
-
-        Ok(Token {
-            address: token_address,
-            symbol: "DUMMY".to_string(),
-            name: "Dummy Token".to_string(),
-            decimals: 18,
-            chain_id: self.chain_id(),
-        })
-    }
-
     async fn get_liquidity_distribution(
         &self,
         pool_address: Address,
