+++ src/dexes/utils.rs
@@ -1,0 +1,23 @@
+use crate::error::Error;
+use crate::models::Token;
+use crate::providers::EthereumProvider;
+use alloy_primitives::Address;
+use std::sync::Arc;
+
+/// Shared implementation of get_token for all DEX protocols
+pub async fn get_token(
+    provider: Arc<EthereumProvider>,
+    token_address: Address,
+    chain_id: u64,
+) -> Result<Token, Error> {
+    // This is a placeholder implementation
+    // In production, we'd use provider.call() with correct parameters to get token info
+
+    Ok(Token {
+        address: token_address,
+        symbol: "DUMMY".to_string(),
+        name: "Dummy Token".to_string(),
+        decimals: 18,
+        chain_id,
+    })
+}
